optimize query in `trace_l7_flows`
1. `worker_numbers` 修改为配置，避免在核心数 < 10 的机器上浪费调度能力
2. 移除 _id to _id_str 逻辑
3. 减少 `construct_from_dataframe` 调用，将 tcp_seq/syscall/x_request_id 的查询条件构造通过 dataframe column list 获取（理论上比逐行迭代快，但数据量少可能不明显）

- remove _id_str 测试结果:
测试集: `spans=18`, 迭代次数 4

| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| before(v6.5.8) | 395 | 1.29 | 2,625ms |
| after - 2 | 477 | 1.56 | 2,028ms |

再增加 3 修改后的测试结果：（吞吐差异不明显，但平均时延有提升）
| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| after - 3 | 478 | 1.56 | 1,985 ms |

另一个测试集：`spans=79`，迭代次数 23

| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| before(v6.5.8) | 75 | 0.24 | 17,792ms |
| after - 2&3 | 133 | 0.43 | 10,006 ms |