@taloric 2024-06-13
对于 dataframe 读数据，小数据量可能差异不是特别明显，写了个简单测试：
https://gist.github.com/taloric/916309861e18e97945fe554f15639523

分别用三种方法读取 dataframe 列数据：
1. `for index in df.index:`
2. `for row in df.itertuples():`
3. `df['column'].tolist()`

在 100000 样本量下三种方法的测试结果是：
```
1. 6.277893781661987s
2. 0.38400983810424805s
3. 0.035607337951660156s
```