@taloric 2024-06-13
4. 将 _id =xx or _id =xxx... 条件修改为 _id IN (xx)，分组条件为 _id >> 32 得到的秒时间戳
对 `spans=18` 数据集，差异不太明显，也不构成负优化

| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| after - 4 | 474 | 1.55 | 2,040 ms |

对 `spans=79`  数据集，也不太明显，可能有点波动

| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| after - 4 | 97 | 0.32 | 13,908 ms |