@taloric 2024-06-13
5. 根据所有 _id 的结果，从中获取 min_time 与 max_time，缩小查询范围，测试差异不大，也不构成负优化

| | Total Request Sent | Request/second | Avg Resp Time |
|--|--|--|--|
| after - 4 | 99 | 0.32 | 13,661 |